// platform/packages/styles/src/types.ts
export type ResponsiveVal<T> =
  | T
  | { default?: T; '2xl'?: T; xl?: T; lg?: T; md?: T; sm?: T; xs?: T }